Обновил систему умений.

+++ Assets/Scripts/Battle/Stats/ElementEfficacy.cs
@@ -6,18 +6,18 @@
 [System.Serializable]
 public class ElementEfficacy
 {
-    [SerializeField] private float _efficacyRate = 100;
+    [SerializeField] private int _efficacyRate = 100;
     [SerializeField] private ElementType _element = ElementType.Fire;
     /// <summary>
     /// Коэффициент эффективности (0 - нет эффекта, 100 - нормальный урон, 200 - урон х2 и т.п.).
     /// </summary>
-    public float EfficacyRate => _efficacyRate;
+    public int EfficacyRate => _efficacyRate;
     /// <summary>
     /// Стихийный элемент.
     /// </summary>
     public ElementType Element => _element;
 
-    public ElementEfficacy(float efficacyRate, ElementType element)
+    public ElementEfficacy(int efficacyRate, ElementType element)
     {
         _efficacyRate = efficacyRate;
         _element = element;
